requirements.txt
requests
pandas
snowflake-connector-python[pandas]
python-dotenv